@@ -1,5 +1,5 @@
 {
-  "name": "acc",
+  "name": "acc-sv5",
   "private": true,
   "version": "0.0.0",
   "type": "module",
@@ -7,24 +7,23 @@
     "dev": "vite",
     "build": "vite build",
     "preview": "vite preview",
-    "check": "svelte-check --tsconfig ./tsconfig.json"
+    "check": "svelte-check --tsconfig ./tsconfig.json && tsc -p tsconfig.node.json"
   },
   "devDependencies": {
-    "@sveltejs/vite-plugin-svelte": "^2.4.2",
-    "@tsconfig/svelte": "^5.0.0",
-    "sass": "^1.63.6",
-    "svelte": "^4.2.1",
-    "svelte-check": "^3.4.4",
-    "svelte-preprocess-cssmodules": "^2.2.4",
-    "tslib": "^2.6.0",
-    "typescript": "^5.0.2",
-    "vite": "^4.4.0"
+    "@sveltejs/vite-plugin-svelte": "^4.0.0",
+    "@tsconfig/svelte": "^5.0.4",
+    "@types/node": "^22.8.6",
+    "sass": "^1.80.5",
+    "svelte": "^5.1.3",
+    "svelte-check": "^4.0.5",
+    "svelte-preprocess-cssmodules": "^2.2.5",
+    "tslib": "^2.8.0",
+    "typescript": "~5.6.2",
+    "vite": "^5.4.10"
   },
   "dependencies": {
-    "@types/node": "^20.4.1",
-    "idb": "^7.1.1",
-    "localbase": "^0.7.5",
-    "nanoid": "^4.0.2",
-    "svelte-spa-router": "^3.3.0"
+    "idb": "^8.0.0",
+    "nanoid": "^5.0.8",
+    "svelte-spa-router": "^4.0.1"
   }
 }
