@@ -1,8 +1,9 @@
+import { mount } from 'svelte';
 import './main.scss';
 import App from './App.svelte';
 
-const app = new App({
-  target: document.getElementById('app')
+const app = mount(App, {
+  target: document.getElementById('app')!
 });
 
 export default app;
